requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile '.\requirements.in'
#
aiohttp==3.8.1
    # via -r .\requirements.in
aiosignal==1.2.0
    # via aiohttp
asgiref==3.5.2
    # via flask
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via aiohttp
beautifulsoup4==4.11.1
    # via -r .\requirements.in
charset-normalizer==2.1.1
    # via aiohttp
click==8.1.3
    # via flask
colorama==0.4.5
    # via click
flask[async]==2.1.2
    # via -r .\requirements.in
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
gunicorn==20.1.0
    # via -r .\requirements.in
idna==3.4
    # via yarl
importlib-metadata==5.0.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
lxml==4.8.0
    # via -r .\requirements.in
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
numpy==1.23.3
    # via pandas
pandas==1.4.2
    # via -r .\requirements.in
pillow==9.2.0
    # via -r .\requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.20.0
    # via -r .\requirements.in
pytz==2022.4
    # via pandas
six==1.16.0
    # via python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
werkzeug==2.2.2
    # via flask
yarl==1.8.1
    # via aiohttp
zipp==3.9.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
